I've modified the ai to now run faster and more effeciently. It now requires you to set up an instance of the ai class for it to work.
I've added methods to the board class to be able to make a move and undo a move when given the poisiton and
any potential piece that was captured.

I've added methods to the piece class to give the piece's value, to give how good a square is for the piece,
and to also output an identifing number for the piece.